[QuestionItem] Fixed text-select causing drag event bug & Combined display/edit in QuestionEdit
