+++ src/pages/FormEdit/QuestionItem/QuestionEdit/Actions/Actions.module.scss
@@ -1,0 +1,17 @@
+.actions {
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+  flex-wrap: wrap;
+}
+
+.action-left {
+  display: flex;
+  align-items: center;
+}
+
+.switch {
+  margin-left: 16px;
+  display: flex;
+  align-items: center;
+}
