+++ src/pages/FormEdit/QuestionItem/QuestionItem.tsx
@@ -1,7 +1,6 @@
 import * as React from 'react';
 import { QuestionField } from '@/types/question';
 import QuestionEdit from './QuestionEdit/QuestionEdit';
-import QuestionDisplay from './QuestionDisplay/QuestionDisplay';
 import { ClickAwayListener } from '@mui/base/ClickAwayListener';
 import DragIndicatorIcon from '@mui/icons-material/DragIndicator';
 
@@ -43,20 +42,22 @@ const QuestionItem = (props: QuestionItemProps) => {
   return (
     <ClickAwayListener mouseEvent="onMouseDown" touchEvent="onTouchStart" onClickAway={onQuestionClickAway}>
       <div
-        className={cx('root', { active, error }, className)}
+        className={cx('root', { editing: active, error }, className)}
         onClick={onClick}
         data-position={index}
-        draggable
+        draggable={!active}
         onDragStart={onDragStart}
         onDragOver={onDragOver}
         onDrop={onDrop}
         onDragLeave={onDragLeave}
         style={style}
       >
-        <div className={cx('drag')}>
-          <DragIndicatorIcon className={cx('drag-icon')} />
-        </div>
-        {active ? <QuestionEdit qId={qId} index={index} /> : <QuestionDisplay qId={qId} index={index} />}
+        {!active && (
+          <div className={cx('drag')}>
+            <DragIndicatorIcon className={cx('drag-icon')} />
+          </div>
+        )}
+        <QuestionEdit qId={qId} index={index} isEditMode={active} />
       </div>
     </ClickAwayListener>
   );
